Organize and divide code for Tab content
