+++ client/src/components/ModalAction.jsx
@@ -4,7 +4,7 @@ import { Card, CardHeader, CardBody, Image, CardFooter, Button, ScrollShadow } f
 // Content Modal
 import ContentWork from "./ContentModal/Work/Work";
 import ContentBank from "./ContentModal/Bank/Bank";
-import ContentGrocery from "./ContentModal/Grocery/Grocery"
+import ContentGrocery from "./ContentModal/Grocery"
 
 export default function ModalAction({ item, onClose }) {
   useEffect(() => {
